# api/app/docs/swagger.yaml
definitions:
...
  hrms_modules_payrollrun_internal_dto.Meta:
    properties:
      currentPage:
        type: integer
      totalItems:
        type: integer
      totalPages:
        type: integer
    type: object
...
  hrms_modules_payrollrun_internal_feature_items_get.GetResponse:
    properties:
      advanceAmount:
        type: number
      advanceDiffAmount:
        type: number
      advanceRepayAmount:
        type: number
      allowDoctorFee:
        type: boolean
      allowElectric:
        type: boolean
      allowHousing:
        type: boolean
      allowInternet:
        type: boolean
      allowWater:
        type: boolean
      attendanceBonusNoLate:
        type: number
      attendanceBonusNoLeave:
        type: number
      bankAccount:
        type: string
      bankName:
        type: string
      bonusAmount:
        type: number
      deductionTotal:
        type: number
      departmentName:
        type: string
      doctorFee:
        type: number
      electricAmount:
        type: number
      electricMeterCurr:
        type: number
      electricMeterPrev:
        type: number
      electricityRatePerUnit:
        type: number
      employeeId:
        type: string
      employeeName:
        type: string
      employeeNumber:
        type: string
      employeeTypeCode:
        type: string
      employeeTypeName:
        type: string
      housingAllowance:
        type: number
      id:
        type: string
      incomeAccumPrev:
        type: number
      incomeAccumTotal:
        type: number
      incomeTotal:
        type: number
      internetAmount:
        type: number
      lateMinutesDeduction:
        type: number
      lateMinutesQty:
        type: integer
      leaveCompensationAmount:
        type: number
      leaveDaysDeduction:
        type: number
      leaveDaysQty:
        type: number
      leaveDoubleDeduction:
        type: number
      leaveDoubleQty:
        type: number
      leaveHoursDeduction:
        type: number
      leaveHoursQty:
        type: number
      loanOutstandingPrev:
        type: number
      loanOutstandingTotal:
        type: number
      loanRepayments:
        items:
          additionalProperties: true
          type: object
        type: array
      netPay:
        type: number
      otAmount:
        type: number
      otHours:
        type: number
      othersDeduction:
        items:
          additionalProperties: true
          type: object
        type: array
      othersIncome:
        items:
          additionalProperties: true
          type: object
        type: array
      pfAccumPrev:
        type: number
      pfAccumTotal:
        type: number
      pfMonthAmount:
        type: number
      photoId:
        type: string
      positionName:
        type: string
      providentFundContribute:
        type: boolean
      ptHourlyRate:
        description: part-time only
        type: number
      ptHoursWorked:
        description: part-time only
        type: number
      runId:
        type: string
      salaryAmount:
        type: number
      ssoAccumPrev:
        type: number
      ssoAccumTotal:
        type: number
      ssoContribute:
        type: boolean
      ssoDeclaredWage:
        type: number
      ssoMonthAmount:
        type: number
      status:
        type: string
      taxAccumPrev:
        type: number
      taxAccumTotal:
        type: number
      taxMonthAmount:
        type: number
      waterAmount:
        type: number
      waterMeterCurr:
        type: number
      waterMeterPrev:
        type: number
      waterRatePerUnit:
        type: number
      withholdTax:
        type: boolean
    type: object
...
  hrms_modules_payrollrun_internal_feature_items_update.UpdateRequest:
    properties:
      advanceRepayAmount:
        type: number
      bonusAmount:
        type: number
      doctorFee:
        type: number
      electricAmount:
        type: number
      electricMeterCurr:
        type: number
      electricMeterPrev:
        type: number
      internetAmount:
        type: number
      lateMinutesDeduction:
        type: number
      lateMinutesQty:
        type: integer
      leaveCompensationAmount:
        type: number
      loanRepayments:
        items:
          additionalProperties: true
          type: object
        type: array
      otAmount:
        type: number
      otHours:
        type: number
      othersDeduction:
        items:
          additionalProperties: true
          type: object
        type: array
      othersIncome:
        items:
          additionalProperties: true
          type: object
        type: array
      pfMonthAmount:
        type: number
      salaryAmount:
        type: number
      ssoMonthAmount:
        type: number
      taxMonthAmount:
        type: number
      waterAmount:
        type: number
      waterMeterCurr:
        type: number
      waterMeterPrev:
        type: number
    type: object
  hrms_modules_payrollrun_internal_feature_items_update.UpdateResponse:
    properties:
      item:
        $ref: '#/definitions/hrms_modules_payrollrun_internal_repository.Item'
    type: object
  hrms_modules_payrollrun_internal_feature_list.Response:
    properties:
      data:
        items:
          $ref: '#/definitions/hrms_modules_payrollrun_internal_dto.Run'
        type: array
      meta:
        $ref: '#/definitions/hrms_modules_payrollrun_internal_dto.Meta'
    type: object
...
  hrms_modules_payrollrun_internal_repository.Item:
    properties:
      advanceAmount:
        type: number
      allowDoctorFee:
        type: boolean
      allowElectric:
        type: boolean
      allowHousing:
        type: boolean
      allowInternet:
        type: boolean
      allowWater:
        type: boolean
      bankAccount:
        type: string
      bankName:
        type: string
      bonusAmount:
        type: number
      deductionTotal:
        type: number
      departmentName:
        type: string
      doctorFee:
        type: number
      employeeID:
        type: string
      employeeName:
        type: string
      employeeNumber:
        type: string
      employeeTypeCode:
        type: string
      employeeTypeName:
        type: string
      id:
        type: string
      incomeAccumPrev:
        type: number
      incomeAccumTotal:
        type: number
      incomeTotal:
        type: number
      lateMinutesDeduction:
        type: number
      lateMinutesQty:
        type: integer
      leaveCompensationAmount:
        type: number
      leaveDaysDeduction:
        type: number
      leaveDaysQty:
        type: number
      loanOutstandingTotal:
        type: number
      netPay:
        type: number
      otAmount:
        type: number
      otHours:
        type: number
      photoID:
        type: string
      positionName:
        type: string
      providentFundContrib:
        type: boolean
      pthourlyRate:
        type: number
      pthoursWorked:
        type: number
      runID:
        type: string
      salaryAmount:
        type: number
      ssoContribute:
        type: boolean
      ssoMonthAmount:
        type: number
      status:
        type: string
      taxMonthAmount:
        type: number
      withholdTax:
        type: boolean
    type: object
...
  hrms_modules_salaryraise_internal_dto.Meta:
    properties:
      currentPage:
        type: integer
      totalItems:
        type: integer
      totalPages:
        type: integer
    type: object
...
  hrms_modules_salaryraise_internal_feature_create.Command:
    properties:
      periodEndDate:
        type: string
      periodStartDate:
        type: string
    required:
    - periodEndDate
    - periodStartDate
    type: object
  hrms_modules_salaryraise_internal_feature_create.Response:
    properties:
      createdAt:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/hrms_modules_salaryraise_internal_dto.Item'
        type: array
      periodEndDate:
        type: string
      periodStartDate:
        type: string
      status:
        type: string
      totalEmployees:
        type: integer
      totalRaiseAmount:
        type: number
      updatedAt:
        type: string
    type: object
...
  hrms_modules_salaryraise_internal_feature_list.Response:
    properties:
      data:
        items:
          $ref: '#/definitions/hrms_modules_salaryraise_internal_dto.Cycle'
        type: array
      meta:
        $ref: '#/definitions/hrms_modules_salaryraise_internal_dto.Meta'
    type: object
...
  payrollrun_internal_feature_create.Command:
    properties:
      payDate:
        type: string
      payrollMonthDate:
        type: string
      periodStartDate:
        type: string
      socialSecurityRateEmployee:
        minimum: 0
        type: number
      socialSecurityRateEmployer:
        minimum: 0
        type: number
    required:
    - payDate
    - payrollMonthDate
    - periodStartDate
    type: object
  payrollrun_internal_feature_create.Response:
    properties:
      approvedAt:
        type: string
      approvedBy:
        type: string
      bonusYear:
        type: integer
      createdAt:
        type: string
      id:
        type: string
      message:
        type: string
      orgProfileSnapshot:
        additionalProperties: true
        type: object
      payDate:
        type: string
      payrollMonthDate:
        type: string
      periodStartDate:
        type: string
      socialSecurityRateEmployee:
        type: number
      socialSecurityRateEmployer:
        type: number
      status:
        type: string
      totalEmployees:
        type: integer
      totalNetPay:
        type: number
      totals:
        $ref: '#/definitions/hrms_modules_payrollrun_internal_dto.RunTotals'
      updatedAt:
        type: string
    type: object
  payrollrun_internal_feature_get.Response:
    properties:
      approvedAt:
        type: string
      approvedBy:
        type: string
      bonusYear:
        type: integer
      createdAt:
        type: string
      id:
        type: string
      orgProfileSnapshot:
        additionalProperties: true
        type: object
      payDate:
        type: string
      payrollMonthDate:
        type: string
      periodStartDate:
        type: string
      socialSecurityRateEmployee:
        type: number
      socialSecurityRateEmployer:
        type: number
      status:
        type: string
      totalEmployees:
        type: integer
      totalNetPay:
        type: number
      totals:
        $ref: '#/definitions/hrms_modules_payrollrun_internal_dto.RunTotals'
      updatedAt:
        type: string
    type: object
...
  payrollrun_internal_feature_items_list.ListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/hrms_modules_payrollrun_internal_dto.Item'
        type: array
      meta:
        $ref: '#/definitions/hrms_modules_payrollrun_internal_dto.Meta'
    type: object
...
  payrollrun_internal_feature_update.Command:
    properties:
      payDate:
        type: string
      status:
        enum:
        - pending
        - approved
        type: string
    type: object
  payrollrun_internal_feature_update.Response:
    properties:
      approvedAt:
        type: string
      approvedBy:
        type: string
      bonusYear:
        type: integer
      createdAt:
        type: string
      id:
        type: string
      message:
        type: string
      orgProfileSnapshot:
        additionalProperties: true
        type: object
      payDate:
        type: string
      payrollMonthDate:
        type: string
      periodStartDate:
        type: string
      socialSecurityRateEmployee:
        type: number
      socialSecurityRateEmployer:
        type: number
      status:
        type: string
      totalEmployees:
        type: integer
      totalNetPay:
        type: number
      totals:
        $ref: '#/definitions/hrms_modules_payrollrun_internal_dto.RunTotals'
      updatedAt:
        type: string
    type: object
...
  salaryraise_internal_feature_get.Response:
    properties:
      createdAt:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/hrms_modules_salaryraise_internal_dto.Item'
        type: array
      periodEndDate:
        type: string
      periodStartDate:
        type: string
      status:
        type: string
      totalEmployees:
        type: integer
      totalRaiseAmount:
        type: number
      updatedAt:
        type: string
    type: object
...
  salaryraise_internal_feature_update.RequestBody:
    properties:
      periodEndDate:
        type: string
      periodStartDate:
        type: string
      status:
        type: string
    type: object
  salaryraise_internal_feature_update.Response:
    properties:
      createdAt:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/hrms_modules_salaryraise_internal_dto.Item'
        type: array
      periodEndDate:
        type: string
      periodStartDate:
        type: string
      status:
        type: string
      totalEmployees:
        type: integer
      totalRaiseAmount:
        type: number
      updatedAt:
        type: string
    type: object
...
info:
  contact: {}
  description: HR Payroll REST API
  title: HR Payroll API
  version: "1.0"
paths:
...
  /payroll-items/{id}:
    get:
      description: ดูรายละเอียดสลิปพนักงาน
      parameters:
      - description: item id
        in: path
        name: id
        required: true
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hrms_modules_payrollrun_internal_feature_items_get.GetResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Get payroll item detail
      tags:
      - Payroll Run
    patch:
      consumes:
      - application/json
      description: แก้ไขยอดในสลิป (เฉพาะรันที่ไม่ approved)
      parameters:
      - description: item id
        in: path
        name: id
        required: true
        type: string
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hrms_modules_payrollrun_internal_feature_items_update.UpdateRequest'
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hrms_modules_payrollrun_internal_feature_items_update.UpdateResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Adjust payroll item
      tags:
      - Payroll Run
  /payroll-runs:
    get:
      description: ดึงประวัติงวดเงินเดือน
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: pending|approved|all
        in: query
        name: status
        type: string
      - description: filter by year of payrollMonthDate
        in: query
        name: year
        type: integer
      - description: YYYY-MM-DD (will use month & year from this date to filter payroll_month_date)
        in: query
        name: monthDate
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hrms_modules_payrollrun_internal_feature_list.Response'
      security:
      - BearerAuth: []
      summary: List payroll runs
      tags:
      - Payroll Run
    post:
      consumes:
      - application/json
      description: สร้างงวดจ่ายเงินเดือน (มีได้ 1 งวดต่อสาขาต่อเดือน)
      parameters:
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payrollrun_internal_feature_create.Command'
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/payrollrun_internal_feature_create.Response'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
      - BearerAuth: []
      summary: Create payroll run
      tags:
      - Payroll Run
  /payroll-runs/{id}:
    delete:
      description: ลบงวดเงินเดือน (เฉพาะ pending)
      parameters:
      - description: run id
        in: path
        name: id
        required: true
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Delete payroll run
      tags:
      - Payroll Run
    get:
      description: ดูข้อมูลงวดเงินเดือน
      parameters:
      - description: run id
        in: path
        name: id
        required: true
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payrollrun_internal_feature_get.Response'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Get payroll run detail
      tags:
      - Payroll Run
    patch:
      consumes:
      - application/json
      description: แก้ payDate หรืออนุมัติ run
      parameters:
      - description: run id
        in: path
        name: id
        required: true
        type: string
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payrollrun_internal_feature_update.Command'
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payrollrun_internal_feature_update.Response'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Update/approve payroll run
      tags:
      - Payroll Run
  /payroll-runs/{id}/items:
    get:
      description: รายการสลิปเงินเดือนใน run
      parameters:
      - description: run id
        in: path
        name: id
        required: true
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: employee name
        in: query
        name: search
        type: string
      - description: employee type code
        in: query
        name: employeeTypeCode
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payrollrun_internal_feature_items_list.ListResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: List payroll items
      tags:
      - Payroll Run
...
  /salary-raise-cycles:
    get:
      description: ดึงรอบปรับเงินเดือน
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: pending|approved|rejected|all
        in: query
        name: status
        type: string
      - description: filter by periodStartDate year
        in: query
        name: year
        type: integer
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hrms_modules_salaryraise_internal_feature_list.Response'
      security:
      - BearerAuth: []
      summary: List salary raise cycles
      tags:
      - Salary Raise
    post:
      consumes:
      - application/json
      description: สร้างรอบปรับเงินเดือน (pending) มีได้ 1 รอบต่อสาขา
      parameters:
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hrms_modules_salaryraise_internal_feature_create.Command'
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hrms_modules_salaryraise_internal_feature_create.Response'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
      security:
      - BearerAuth: []
      summary: Create salary raise cycle
      tags:
      - Salary Raise
  /salary-raise-cycles/{id}:
    delete:
      parameters:
      - description: cycle id
        in: path
        name: id
        required: true
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Delete salary raise cycle (soft delete)
      tags:
      - Salary Raise
    get:
      description: ดูรอบปรับเงินเดือน (รวม items)
      parameters:
      - description: cycle id
        in: path
        name: id
        required: true
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/salaryraise_internal_feature_get.Response'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Get salary raise cycle detail
      tags:
      - Salary Raise
    patch:
      consumes:
      - application/json
      description: แก้ไขช่วงเวลา หรือเปลี่ยนสถานะ (Approve/Reject)
      parameters:
      - description: cycle id
        in: path
        name: id
        required: true
        type: string
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/salaryraise_internal_feature_update.RequestBody'
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/salaryraise_internal_feature_update.Response'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Update salary raise cycle
      tags:
      - Salary Raise
...
securityDefinitions:
  BearerAuth:
    description: Add "Bearer <access token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"