# api/modules/payrollrun/internal/feature/items/get/endpoint.go
package itemsget

import (
	"github.com/gofiber/fiber/v3"
	"github.com/google/uuid"

	"hrms/shared/common/errs"
	"hrms/shared/common/mediator"
	"hrms/shared/common/response"
)

// @Summary Get payroll item detail
// @Description ดูรายละเอียดสลิปพนักงาน
// @Tags Payroll Run
// @Produce json
// @Security BearerAuth
// @Param id path string true "item id"
// @Success 200 {object} GetResponse
// @Failure 400
// @Failure 401
// @Failure 403
// @Failure 404
// @Param X-Company-ID header string false "Company ID"
// @Param X-Branch-ID header string false "Branch ID"
// @Router /payroll-items/{id} [get]
func NewEndpoint(router fiber.Router) {
	router.Get("/:itemId", func(c fiber.Ctx) error {
		itemID, err := uuid.Parse(c.Params("itemId"))
		if err != nil {
			return errs.BadRequest("invalid item id")
		}
		resp, err := mediator.Send[*GetQuery, *GetResponse](c.Context(), &GetQuery{
			ID: itemID,
		})
		if err != nil {
			return err
		}
		return response.JSON(c, fiber.StatusOK, resp.ItemDetail)
	})
}